typed subscription decorator and supporting types moved to separate files
