+++ src/common/decorators/index.ts
@@ -1,4 +1,4 @@
 export * from './current-user.decorator';
 export * from './profiling.decorator';
 export * from './authorization.agent.privilege';
-export * from './subscription.decorator';
+export * from './typed.subscription';
