+++ src/common/decorators/typed.subscription/subscription.resolve.context.ts
@@ -1,0 +1,10 @@
+import { IncomingHttpHeaders } from 'http';
+import { AgentInfo } from '@core/authentication';
+
+export interface SubscriptionResolveContext {
+  req: {
+    authInfo: unknown;
+    headers: IncomingHttpHeaders;
+    user: AgentInfo;
+  };
+}
