+++ src/domain/collaboration/callout/callout.resolver.subscriptions.ts
@@ -16,7 +16,7 @@ import {
 import { CalloutService } from '@domain/collaboration/callout/callout.service';
 import { CalloutAspectCreated } from '@domain/collaboration/callout';
 import { UUID } from '@domain/common/scalars';
-import { TypedSubscription } from '@common/decorators/subscription.decorator';
+import { TypedSubscription } from '@common/decorators/typed.subscription/typed.subscription.decorator';
 import { CalloutMessageReceivedArgs } from './dto/callout.message.received.args';
 import { CalloutMessageReceived } from './dto/callout.dto.event.message.received';
 import { CalloutAspectCreatedArgs } from './dto/callout.aspect.created.args';
