+++ src/domain/communication/communication/dto/communication.discussion.updated.args.ts
@@ -4,7 +4,7 @@ import { UUID } from '@domain/common/scalars';
 @ArgsType()
 export class CommunicationDiscussionUpdatedArgs {
   @Field(() => UUID, {
-    description: 'The Communication to receive Updates from.',
+    description: 'The ID of the Communication to subscribe to all updates on.',
     nullable: false,
   })
   communicationID!: string;
